Check if there are keyframes before deleting them inside: def private_prm_deleteAllKeyframes(...)

+++ src/py__3_11/py_flam3__3_11_H21_UP.py
@@ -2488,9 +2488,10 @@ class flam3h_general_utils
         """ 
         if isinstance(_prm, str): prm: hou.Parm = node.parm(_prm)
         elif isinstance(_prm, hou.Parm): prm: hou.Parm = _prm
-        prm.lock(False)
-        prm.deleteAllKeyframes()
-        prm.lock(True)
+        if len(prm.keyframes()):
+            prm.lock(False)
+            prm.deleteAllKeyframes()
+            prm.lock(True)
         
         
     @staticmethod
